Rename `bloodpressures.view` package to `bloodpressures.ui`
