+++ app/src/main/res/layout/screen_patient_summary.xml
@@ -75,7 +75,7 @@
         android:layout_marginTop="@dimen/spacing_8"
         android:layout_marginEnd="@dimen/spacing_8" />
 
-      <org.simple.clinic.summary.bloodpressures.view.BloodPressureSummaryView
+      <org.simple.clinic.summary.bloodpressures.ui.BloodPressureSummaryView
         android:id="@+id/bloodPressureSummaryView"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
